chain: split serialization.rs into files
The code is unchanged except for removing an unneeded clippy annotation in proptests.rs.
